Increased the version of jest on which this package depends.
jest depends on jest-cli, which depends on jest-haste-map, which depends on sane, which depends on an older version of exec-sh, which depends on merge, and a vulnerability was found in merge. (See https://hackerone.com/reports/381194) Depending on a newer version of jest brings in a newer version of merge, in which this seems to be resolved. Note that since we're only using jest for testing, this vulnerability never really affected this package.