Made validation slightly stricter.
The previous version would accept any three-character string where only the first is upper case as the day of the week (such as "Abc"). The new version is slightly smarter, as it only allows a subset of characters. For instance, it does not allow the day of the week to start with A, B, C, D, E, X, Y, or Z.